chore(LESQ-2007): move pages router pages out of the way
were replacing them with app router equivalents
