+++ src/__tests__/pages/teachers/beta/lessons/[lessonSlug]/media.test.tsx
@@ -7,7 +7,7 @@ import {
   getStaticPaths,
   URLParams,
 } from "@/pages/teachers/beta/lessons/[lessonSlug]/media";
-import { LessonMediaClipsPage } from "@/pages/teachers/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug]/media";
+import { LessonMediaClipsPage } from "@/pages/teachers/__legacy__programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug]/media";
 import curriculumApi from "@/node-lib/curriculum-api-2023/__mocks__";
 import lessonMediaClipsFixtures from "@/node-lib/curriculum-api-2023/fixtures/lessonMediaClips.fixture";
 import { CanonicalLessonMediaClipsPageProps } from "@/pages/teachers/lessons/[lessonSlug]/media";
